Version 0.01.006.0004 - Add unit tests. Passes all unit tests.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/aarre/src/NonNegativeInteger.java
@@ -29,6 +29,6 @@ public class NonNegativeInteger {
 	}
 
 	public String toString() {
-		return String.format("%f", this.value);
+		return String.format("%d", this.value.intValue());
 	}
 }
